+++ .travis.yml
@@ -1,0 +1,9 @@
+lanague: java
+
+jdk: openjdk8
+
+install:
+  - mvn clean install
+
+after_success:
+  - echo "Success!"
