--- src/main/java/io/viro/p2pfs/telnet/P2PFSMessageParser.java
@@ -1,17 +1,0 @@
-package io.viro.p2pfs.telnet;
-
-/**
- * Parse messages.
- */
-public class P2PFSMessageParser {
-
-    /**
-     * Returns a sort of command. Look up the commands in the helping code.(chanuka)
-     *
-     * @param message
-     * @return
-     */
-    public String parseMessage(String message) {
-        return "None";
-    }
-}
