+++ src/main/java/io/viro/p2pfs/telnet/message/Response.java
@@ -1,0 +1,4 @@
+package io.viro.p2pfs.telnet.message;
+
+public class Response {
+}
